Revert "Merge branch 'Fix/nifi-integration-timeout' into feature/nifi-integration-record-listener"
This reverts commit ef4d9dec4b41347fcd296da53d4ce6cb3513e586, reversing
changes made to 3dfa7ba1f86c2ccf059b2781395732d80e1dd440.

+++ plc4j/integrations/apache-nifi/nifi-plc4x-processors/src/main/java/org/apache/plc4x/nifi/BasePlc4xProcessor.java
@@ -58,7 +58,7 @@ public abstract class BasePlc4xProcessor extends AbstractProcessor {
   
     protected String connectionString;
     protected Map<String, String> addressMap;
-    protected Long timeout;
+    protected Integer timeout;
 
     protected final SchemaCache schemaCache = new SchemaCache(0);
 
@@ -170,7 +170,7 @@ public abstract class BasePlc4xProcessor extends AbstractProcessor {
 		connectionString = context.getProperty(PLC_CONNECTION_STRING.getName()).getValue();
         schemaCache.restartCache(context.getProperty(PLC_SCHEMA_CACHE_SIZE).asInteger());
         debugEnabled = getLogger().isDebugEnabled();
-        timeout = Long.valueOf(context.getProperty(PLC_FUTURE_TIMEOUT_MILISECONDS.getName()).getValue());
+        timeout = Integer.valueOf(context.getProperty(PLC_FUTURE_TIMEOUT_MILISECONDS.getName()).getValue());
     }
 
     @Override
